fix image pull from peoxy issue

+++ scripts/prepare-env.sh
@@ -200,7 +200,10 @@ shift $((OPTIND-1))
 
 for profile in lb k8s k8s-cloud-init k8s-cloud-init-local-registries;
 do
-  lxc profile delete "$profile"
+  exists=$(lxc profile ls | grep "$profile")
+  if [ "$exists" != "" ]; then
+    lxc profile delete "$profile"
+  fi
 done
 
 cat <<EOF | sudo lxd init --preseed
@@ -691,7 +694,7 @@ image=$(lxc image ls | grep focal-cloud)
 if [ "$image" == "" ]; then
   if [ "$slim" == "" ]; then
     VERSION=$(curl -sSL https://cloud-images.ubuntu.com/daily/streams/v1/com.ubuntu.cloud:daily:download.json | jq '.products."com.ubuntu.cloud.daily:server:20.04:amd64".versions | keys[]' | sort -r | head -1 | tr -d '"')
-    PROXY=$(grep Proxy /etc/apt/apt.conf.d/* | awk '{print $2}' | tr -d ';|"' | sed 's@^http://\(.*\):3142/@\1@')
+    PROXY=$(grep Proxy /etc/apt/apt.conf.d/* | awk '{print $2}' | tr -d ';|"' | sed 's@^\(http://.*:3142\)/@\1@')
     if [ "$PROXY" != "" ]; then
       SERVER=$PROXY
     else
